+++ meteor-app/imports/api/products/server/publications.js
@@ -1,6 +1,7 @@
 import { Meteor } from 'meteor/meteor'
 import Products, { ProductTypes, Carts } from '../schema'
-import Members from '../../members/schema'
+import Members from '/imports/api/members/schema'
+import '../methods'
 
 Meteor.publish('all.products', () => Products.find({}))
 
@@ -15,19 +16,3 @@ Meteor.publish('product.bycode', code => {
 })
 // TODO: check if it is real
 Meteor.publish('cart', (id, memberId) => [Carts.find(id), Members.find(memberId)])
-
-Meteor.methods({
-  'rm.Products': id => {
-    Products.remove(id)
-  },
-  'update.Products': form => {
-    const id = form._id
-    delete form._id
-    Products.update(id, { $set: form })
-  },
-  'insert.Products': form => {
-    console.log('Inserting product', form)
-    const id = Products.insert(form)
-    console.log('Added ', id)
-  }
-})